requirements.txt
crewai-tools
python-dotenv
python-epo-ops-client
streamlit>=1.28.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
requests>=2.28.0
python-dotenv>=1.0.0
crewai>=0.28.0
pydantic>=2.0.0
reportlab>=4.0.0
fpdf2>=2.7.0
plotly>=5.15.0
certifi>=2022.12.7


